requirements.txt
alembic==1.4.3
appdirs==1.4.4
asgiref==3.2.10
bcrypt==3.2.0
beautifulsoup4==4.9.1
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
distlib==0.3.1
Django==3.1.1
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==19.9.0
idna==2.10
importlib-metadata==2.1.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
MouseInfo==0.1.3
Pillow==7.2.0
pipenv==2018.11.26
psycopg2==2.8.6
PyAutoGUI==0.9.50
pycparser==2.20
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-editor==1.0.4
PyTweening==1.0.3
pytz==2020.1
pywhatkit==2.7
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
svgwrite==1.4
Tree==0.2.4
urllib3==1.25.10
virtualenv==20.0.33
virtualenv-clone==0.5.4
visitor==0.1.3
Werkzeug==1.0.1
wikipedia==1.4.0
WTForms==2.3.3
zipp==3.8.1
